backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
asyncpg==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.1
google-generativeai==0.7.2
